Initialize meta boxes whether or not they’re visible
